Added ability to build under Athlon64 CPU architecture.

+++ cpp/tests/test-hashtable.cpp
@@ -36,9 +36,9 @@ struct KeyAsHash
 
 struct AddressAsHash
 {
-   int operator()(const char* key) const
+   long operator()(const char* key) const
    {
-      return (int)key;
+      return (long)key;
    }
 };
